refactor: renaming folders referring to tables
