+++ backend/src/pizzaSizes/pizzaSize.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common'
-import { PizzaSizeService } from 'src/pizzaSize/pizzaSize.service'
+import { PizzaSizeService } from 'src/pizzaSizes/pizzaSize.service'
 
 @Controller('api')
 export class ControllerPizzaSize {
